Reduce chance of very long exploration for region partitioning
The recursive backtrack algorithm eventually finds a solution for
the partitioning, but it can sometimes take very long time to finish.
Probably we should use annealing or such stochastic approach to
bail out from those bad conditions, but for now option selection
optimization seems to work to some extent.

Also, small regions are prioritized, so unsolvable problems are less
likely to be generated.