Recreate the whole operator partitions if solvability check fails
Previously, operator partition generation has backtracked to the nearest
candidate if solvability check fails.  In this logic, if there's a bad
partition in the very early stage of recursive calls (such that the first
and the second rows are the same), it will take very long time to
generate solvable problem because we have to enumerate every
possible permutations.

Starting everything all over may not be able to permute every possible
cases, but can avoid being trapped in a very bad condition.