Increase maximum size to 9 and add a white color for operator partitions
The algorithm has no intrinsic limit in size, but practically 9 would be
the maximum for memos.

I haven't tested if 7 or more sizes can be solved, but the same algorithm
produced solvable puzzles up to 6, so it's likely.

Also, larger sizes tend to take more time to generate a problem (and try
solving), so slow devices may seem frozen.